Move error handler tests to test_route_handlers.py per code review
